(PC-26274)[PRO] refactor: unify abstraction levels
