+++ pro/src/pages/Home/ProfileAndSupport/ProfileAndSupport.module.scss
@@ -1,31 +1,10 @@
 @use "styles/variables/_colors.scss" as colors;
 @use "styles/mixins/_rem.scss" as rem;
 @use "styles/mixins/_fonts.scss" as fonts;
+@use "styles/variables/_size.scss" as size;
 
 .title {
   @include fonts.title3;
 
   margin-bottom: rem.torem(16px);
-}
-
-.description-list {
-  margin-top: rem.torem(16px);
-}
-
-.description-item {
-  margin-top: rem.torem(8px);
-  display: flex;
-  flex-direction: row;
-  gap: rem.torem(8px);
-}
-
-.description-label {
-  color: colors.$grey-dark;
-  flex-shrink: 0;
-}
-
-.description-value {
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
 }
