--- pro/src/pages/Home/ProfileAndSupport/index.ts
@@ -1,1 +1,0 @@
-export { default as ProfileAndSupport } from './ProfileAndSupport'
